a(possibly) better renderer.
i was digging around with the directXTK when i came upon this sample
code, it looks much cleaner than anything mould gave us, it doesn't
track position ,but that's a quick fix. think its worth looking into, or
should we stick with what we have? im not sure where the shaders are
exactly though